Moves getParamStringArray to its own util file
