+++ src/util/clipSketchToGeography.ts
@@ -16,32 +16,7 @@ import {
   getFlatGeobufFilename,
 } from "@seasketch/geoprocessing";
 import simplify from "@turf/simplify";
-import { ExtraParams } from "../types";
 import { getGeographyById } from "./getGeographyById";
-
-/**
- * Validates and returns string[] parameter from extraParams
- * @param param string name of parameter to extract from extraParams
- * @param extraParams parameters passed from client to GP function
- * @returns string[]
- */
-export const getParamStringArray = (
-  param: string,
-  extraParams: ExtraParams
-): string[] => {
-  const value = extraParams[param as keyof ExtraParams];
-  if (!Array.isArray(value)) {
-    throw new Error(
-      `${param} is not an array of strings in ExtraParams -- ${value}`
-    );
-  } else if (!value.length) {
-    throw new Error(`${param} is empty in ExtraParams`);
-  } else if (typeof value[0] !== "string") {
-    throw new Error(
-      `${param} array does not contain strings as expected in ExtraParams`
-    );
-  } else return value;
-};
 
 /**
  * Clips sketch to geography. If geographyId and simplifyOptions are both null, returns original sketch
